moving all controller unittest funcs to suite_test.go (#574)
